💩 Fix cpbin script

+++ cpbin.js
@@ -1,8 +1,6 @@
 const { join } = require('path');
 const { copyFileSync, mkdirSync, existsSync } = require('fs');
 
+const binPath = join(__dirname, 'native', 'index.node');
 if (!existsSync('dist')) mkdirSync('dist');
-copyFileSync(
-  join(__dirname, 'native', 'index.node'),
-  join(__dirname, 'dist', 'index.node'),
-);
+if (!existsSync(join(binPath))) copyFileSync(binPath, join(__dirname, 'dist', 'index.node'));
